increase resilience with consul, fix dedupe checks
Increase interval and service health check timeouts.  This will reduce a
small amount of traffic, intended to be slightly more inclusive of
runners.  Runners which are slow to start can sometimes fail to register
themselves with consul, or fail it to respond to health checks fast
enough while booting when they do register.  In theory, this adjusts the timing to
be slightly more permissive.

In the process of testing, found that Address refers to node details
while ServiceAddress is the correct field to test for equality when
deduping services.  This helps reduce churn in consul when an
initializing dispatcher iterates through all tenants and restarts them.