keep demuxers alive
If a cluster:runner has registered running services, but for some reason
is missing an entry for itself as cluster.service.consul, there will be
a found tenant, but no candidates.
This change sends a 404, which succesfully staves off re-assigning a tenant to a different cluster.  This issues a 404 instead of crashing.