validate correct IP and port mapping
If there are multiple entries in consul for a given Nightscout instance, this
change ensures that the IP and port mapping were meant to be used together.
Prior to this change, nginx would have been allowed to pick any IP, but using
only the selected port which would result in incorrect or mismatched data.

This change ensures that when a port is selected the IP mapping that matches
the port is used correctly instead of allowing NGINX to guess incorrectly.
This fixes an issue that can arise that presents people with the wrong data
from other sites.

Many thanks to Marion Barker for helping to identify this issue.

With this patch, the resolver service performs an additional dns lookup on the node name for the given service which resolves to an IP.  The entire valid upstream backend is provided in a well formed url including the ip and port using the x-selected-backend header.  NGINX then uses the result for the proxy configuration.
While consul's HTTP API provides a node lookup that provides all of the
information required in a single API call, we theorize that two dns requests
are still better performance than a single HTTP request.