force candidate health checks when no 1:1 match is available
This change simplifies the logic during selecting tenant assignments to
runners.
If there is a 1:1 match with a tenant and runner existing assignment, re-use
that assignment.
In the case where there is a tenant but no candidates for our set of runners, force a healthcheck on all available runners.
In the case where there are multiple tenants, find all the matching runners available.