stricter tenant and candidate matching
In order for existing tenant and candidate matches to be made, the existing
tenant service address must also match one of the existing candidate addresses.
In the case where this is not true, ensure that the healthchecks and scheduling
policies are applied for a new assignment.

This is more likely to occur in duplicated environments such as those in
blue-green deployments.