enable publishing additional tenant tag
This change allows finding all tenants as well as tenants residing on a group
of runners.  This is very common in blue/green environmenets.  For example
consider the following DNS queries enabled by this change:
  * tenant.backends.service.consul - all possible tenants
  * cluster.backends.service.consul - all tenants hosted on faux-blue
  * cluster-blue.backends.service.consul - all tenants hosted on blue runners
  * cluster-green.backends.service.consul - all tenants hosted on green runners

This does not affect the usualy lookup for specifically identified tenants
through their WEB_NAME tag.
  * $tenant.backends.service.consul

This allows tenant-availability-keeper and other programs to more easily filter
only relevant tenants when needed.