keep children that are not subject of the active phase
Any children sent to our webhook by metacontroller but absent from the response
will be deleted. There are several example of metacontroller hooks online that
pass the children that the business logic is not actively concerned with
through to the responses.  Doing this allows the objects from prior stages to
be kept indefinitely for operational awareness instead of implicitly deleted.