set writable group permission on socket
In environments like kubernetes, two containers running in the same pod can
communicate using unix domain sockets, but only if the file permissions permit
for each process to independently read and write to the file.
Kubernetes can trivially arrange for different processes/containers to run as
different groups, but the socket must be group writable for both processes to
use it, which is what this change facilitates.