register multienv-deployment-controller in consul
Investigation into causes for DNS ETIMEOUT in the healthchecks and resolvers
suggests overflowing volumes of UDP traffic.  It turns out a significant amount
of traffic is caused by consul agent's healthcheck system.  Consul Agent will
use the underlying OS's resolv.conf to resolve any names consumed by the health
checks.

Each tenant is configured with a healthcheck unique to their service details.
The dns name of this healthcheck service appears to be the source of 70% of
non-tenant traffic, and was set to the name of a kubernetes service backing the
deployment.  This change makes the healthcheck service name a domain of consul
itself, including a terminal dot to indicate that dns search paths should not
be queried, even without a local dns cache in place for consul agent.

This change also allows the service to register itself in consul as
multienv-deployment-controller.service.consul by default.